Add new strategy documents and update existing ones for fingerprint and wallet analysis
- Introduced `fp-4sol-launcher.md` detailing the behavior and metrics of the 4 SOL launcher fingerprint.
- Added `maxbuy-launcher-bundle-band.md` explaining the edge of the max-buy launcher's creation bundle.
- Updated `maxbuy-launcher-fingerprint.md` to clarify the impact of funding-tier fingerprints and their relation to the creation-slot bundle.
- Enhanced `metrics-reference.md` with insights on the relationship between windowed gross flow and lifetime metrics.
- Created `wallet-8dtx-logic.md` to document the reconstructed dip-turn logic for wallet `8dtx2tr4`.
- Revised `wallet-analysis.md` to summarize findings from the wallet analysis and link to the new wallet logic document.

+++ hunter/core/src/services/veteran_roster.rs
@@ -98,10 +98,13 @@ pub async fn launch_history(
     // window is its online approximation (r = 0.999 on the cohort this was derived on).
     let rows = sqlx::query(
         r#"
+        -- `SUM(bigint)` is NUMERIC in Postgres, which does not decode as `i64` - the
+        -- cast is what keeps this readable as one. Safe: these are per-(mint, wallet)
+        -- creation-slot totals, orders of magnitude below the i64 ceiling.
         SELECT tr.mint_address,
-               w.address                                                      AS wallet,
-               SUM(tr.amount_lamports) FILTER (WHERE tr.trade_type = 'buy')   AS buy_lamports,
-               SUM(tr.amount_lamports) FILTER (WHERE tr.trade_type = 'sell')  AS sell_lamports
+               w.address                                                              AS wallet,
+               SUM(tr.amount_lamports) FILTER (WHERE tr.trade_type = 'buy')::bigint   AS buy_lamports,
+               SUM(tr.amount_lamports) FILTER (WHERE tr.trade_type = 'sell')::bigint  AS sell_lamports
           FROM trades tr
           JOIN tokens  tk ON tk.mint_address = tr.mint_address
           JOIN wallet_dict w ON w.id = tr.wallet_id
